requirements.txt
annotated-types==0.6.0
anyio==4.3.0
black==23.7.0
click==8.1.7
distro==1.9.0
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
markdown-it-py==3.0.0
marko==2.0.3
mdurl==0.1.2
mypy==1.9.0
mypy-extensions==1.0.0
openai==1.12.0
pathspec==0.12.1
portkey-ai==1.2.3
prompt-toolkit==3.0.36
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
questionary==2.0.1
requests~=2.31.0
rich==13.7.1
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
uvicorn==0.29.0